requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose
python-multipart
python-dotenv
passlib[bcrypt]